perf(anthropic): defer auth loading until hook execution
