+++ extensions/anthropic/lazy-import.test.ts
@@ -1,6 +1,7 @@
 import type {
   OpenClawPluginNodeHostCommand,
   OpenClawPluginNodeInvokePolicy,
+  ProviderPlugin,
 } from "openclaw/plugin-sdk/plugin-entry";
 import { createTestPluginApi } from "openclaw/plugin-sdk/plugin-test-api";
 import { createPluginRuntimeMock } from "openclaw/plugin-sdk/plugin-test-runtime";
@@ -11,6 +12,7 @@ describe("anthropic session catalog lazy imports", () => {
   afterEach(() => {
     vi.doUnmock("./session-catalog.js");
     vi.doUnmock("./session-catalog-node-commands.js");
+    vi.doUnmock("openclaw/plugin-sdk/provider-auth");
     vi.resetModules();
   });
 
@@ -68,5 +70,40 @@ describe("anthropic session catalog lazy imports", () => {
     await expect(nodeCommands[0]?.handle()).resolves.toBe("[]");
     expect(catalogImports).toBe(1);
     expect(nodeCommandImports).toBe(1);
+  });
+
+  it("registers provider hooks before loading auth execution", async () => {
+    vi.doMock("openclaw/plugin-sdk/provider-auth", () => {
+      throw new Error("auth execution loaded");
+    });
+    const { default: anthropicPlugin } = await import("./index.js");
+    const providers: ProviderPlugin[] = [];
+    anthropicPlugin.register(
+      createTestPluginApi({
+        id: "anthropic",
+        name: "Anthropic",
+        source: "test",
+        config: {},
+        runtime: createPluginRuntimeMock(),
+        registerProvider: (provider) => providers.push(provider),
+      }),
+    );
+    const provider = providers[0]!;
+    expect(provider.auth.map(({ id }) => id)).toEqual(["cli", "setup-token", "api-key"]);
+    expect(
+      provider.classifyFailoverReason?.({
+        provider: "anthropic",
+        errorMessage: "",
+        code: "API_ERROR",
+      }),
+    ).toBe("server_error");
+    await expect(
+      provider.buildAuthDoctorHint!({
+        provider: "anthropic",
+        store: { version: 1, profiles: {} },
+      }),
+    ).rejects.toMatchObject({
+      cause: { message: "auth execution loaded" },
+    });
   });
 });
